Move to lib crate for better organization :godmode:
